requirements.txt
allure-pytest~=2.9.42
allure-python-commons~=2.9.42
pytest~=6.2.4
requests~=2.25.1
python-dotenv~=0.17.1
selenium~=3.141.0